Refactored, clean-up and used Google cod-style
