+++ src/main/java/com/chat/yourway/config/RedisConfig.java
@@ -1,6 +1,6 @@
 package com.chat.yourway.config;
 
-import com.chat.yourway.model.Token;
+import com.chat.yourway.model.token.Token;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -10,38 +10,32 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
-/**
- * {@link RedisConfig}
- *
- * @author Dmytro Trotsenko on 8/1/23
- */
-
 @Configuration
 public class RedisConfig {
 
-    @Value("${spring.data.redis.host}")
-    private String redisHost;
+  @Value("${spring.data.redis.host}")
+  private String redisHost;
 
-    @Value("${spring.data.redis.port}")
-    private int redisPort;
+  @Value("${spring.data.redis.port}")
+  private int redisPort;
 
-    @Value("${spring.data.redis.password}")
-    private String redisPassword;
+  @Value("${spring.data.redis.password}")
+  private String redisPassword;
 
-    @Bean
-    public JedisConnectionFactory jedisConnectionFactory() {
-        var redisStandaloneConfiguration = new RedisStandaloneConfiguration(redisHost, redisPort);
-        redisStandaloneConfiguration.setPassword(redisPassword);
-        return new JedisConnectionFactory(redisStandaloneConfiguration);
-    }
+  @Bean
+  public JedisConnectionFactory jedisConnectionFactory() {
+    var redisStandaloneConfiguration = new RedisStandaloneConfiguration(redisHost, redisPort);
+    redisStandaloneConfiguration.setPassword(redisPassword);
+    return new JedisConnectionFactory(redisStandaloneConfiguration);
+  }
 
-    @Bean
-    public RedisTemplate<String, Token> redisTemplate() {
-        RedisTemplate<String, Token> template = new RedisTemplate<>();
-        template.setConnectionFactory(jedisConnectionFactory());
-        template.setKeySerializer(new StringRedisSerializer());
-        template.setValueSerializer(new GenericJackson2JsonRedisSerializer());
-        return template;
-    }
+  @Bean
+  public RedisTemplate<String, Token> redisTemplate() {
+    RedisTemplate<String, Token> template = new RedisTemplate<>();
+    template.setConnectionFactory(jedisConnectionFactory());
+    template.setKeySerializer(new StringRedisSerializer());
+    template.setValueSerializer(new GenericJackson2JsonRedisSerializer());
+    return template;
+  }
 
 }
