--- src/main/java/com/chat/yourway/config/SecurityConfig.java
@@ -1,52 +1,0 @@
-package com.chat.yourway.config;
-
-import com.chat.yourway.repository.ContactRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-/**
- * {@link SecurityConfig}
- *
- * @author Dmytro Trotsenko on 7/26/23
- */
-
-@Configuration
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-    private final ContactRepository contactRepository;
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        return username -> contactRepository.findByEmail(username)
-                .orElseThrow(() -> new UsernameNotFoundException("Contact not found"));
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        var authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(userDetailsService());
-        authProvider.setPasswordEncoder(passwordEncoder());
-        return authProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authConfig) throws Exception {
-        return authConfig.getAuthenticationManager();
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-}
