+++ src/main/java/com/chat/yourway/controller/ChatController.java
@@ -1,8 +1,8 @@
 package com.chat.yourway.controller;
 
-import com.chat.yourway.dto.request.ReceivedMessage;
-import com.chat.yourway.dto.response.SendMessage;
-import com.chat.yourway.service.MessageService;
+import com.chat.yourway.dto.request.ReceivedMessageDto;
+import com.chat.yourway.dto.response.MessageResponseDto;
+import com.chat.yourway.service.ChatMessageServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
@@ -11,29 +11,23 @@ import org.springframework.stereotype.Controller;
 
 import java.security.Principal;
 
-/**
- * {@link ChatController}
- *
- * @author Dmytro Trotsenko on 7/21/23
- */
-
 @Controller
 @RequiredArgsConstructor
 public class ChatController {
 
-    private final MessageService messageService;
+  private final ChatMessageServiceImpl chatMessageServiceImpl;
 
-    @MessageMapping("/application")
-    @SendTo("/topic")
-    public SendMessage sendToTopic(ReceivedMessage message, Principal principal) {
-        String username = principal.getName();
-        return messageService.sendToTopic(message, username);
-    }
+  @MessageMapping("/application")
+  @SendTo("/topic")
+  public MessageResponseDto sendToTopic(ReceivedMessageDto message, Principal principal) {
+    String username = principal.getName();
+    return chatMessageServiceImpl.sendToTopic(message, username);
+  }
 
-    @MessageMapping("/private")
-    public SendMessage sendToUser(@Payload ReceivedMessage message, Principal principal) {
-        String username = principal.getName();
-        return messageService.sendToUser(message, username);
-    }
+  @MessageMapping("/private")
+  public MessageResponseDto sendToUser(@Payload ReceivedMessageDto message, Principal principal) {
+    String username = principal.getName();
+    return chatMessageServiceImpl.sendToUser(message, username);
+  }
 
 }
