--- src/main/java/com/chat/yourway/controller/DemoController.java
@@ -1,22 +1,0 @@
-package com.chat.yourway.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-/**
- * {@link DemoController}
- *
- * @author Dmytro Trotsenko on 7/27/23
- */
-@RestController
-@RequestMapping("/demo-controller")
-public class DemoController {
-
-    @GetMapping
-    public ResponseEntity<String> sayHello() {
-        return ResponseEntity.ok("Hello from secured endpoint");
-    }
-
-}
