+++ src/main/java/com/chat/yourway/dto/response/AuthResponseDto.java
@@ -5,19 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/**
- * {@link AuthResponseDto}
- *
- * @author Dmytro Trotsenko on 7/26/23
- */
-
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class AuthResponseDto {
 
-    private String accessToken;
-    private String refreshToken;
+  private String accessToken;
+  private String refreshToken;
 
 }
