+++ src/main/java/com/chat/yourway/dto/response/MessageResponseDto.java
@@ -1,0 +1,15 @@
+package com.chat.yourway.dto.response;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class MessageResponseDto {
+
+  private LocalDateTime sentTime;
+  private String sentFrom;
+  private String sendTo;
+  private String text;
+
+}
