+++ src/main/java/com/chat/yourway/model/Role.java
@@ -1,11 +1,6 @@
 package com.chat.yourway.model;
 
-/**
- * {@link Role}
- *
- * @author Dmytro Trotsenko on 7/25/23
- */
 public enum Role {
-    USER,
-    ADMIN
+  USER,
+  ADMIN
 }
