--- src/main/java/com/chat/yourway/model/TokenType.java
@@ -1,10 +1,0 @@
-package com.chat.yourway.model;
-
-/**
- * {@link TokenType}
- *
- * @author Dmytro Trotsenko on 7/28/23
- */
-public enum TokenType {
-    BEARER
-}
