+++ src/main/java/com/chat/yourway/model/email/EmailSend.java
@@ -1,4 +1,4 @@
-package com.chat.yourway.model;
+package com.chat.yourway.model.email;
 
 public record EmailSend(String to,
                         String subject,
