+++ src/main/java/com/chat/yourway/model/token/TokenType.java
@@ -1,0 +1,5 @@
+package com.chat.yourway.model.token;
+
+public enum TokenType {
+  BEARER
+}
