+++ src/main/java/com/chat/yourway/repository/EmailTokenRepository.java
@@ -1,6 +1,6 @@
 package com.chat.yourway.repository;
 
-import com.chat.yourway.model.EmailToken;
+import com.chat.yourway.model.email.EmailToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
